requirements.txt
Django==1.7.4
Pygments==2.0.2
SQLAlchemy==0.9.8
Whoosh==2.5.7
apache-log-parser==1.4.1
bpython==0.13.2
django-haystack==2.3.1
django-jsonview==0.4.3
django-nap==0.14.2
futures==2.2.0
geoip2==2.1.0
grako==3.4.3
gunicorn==19.2
maxminddb==1.1.1
pygeoip==0.3.2
requests==2.5.1
six==1.9.0
sure==1.2.9
